+++ js/utils/index.js
@@ -1,2 +1,5 @@
 export * from './common'
 export * from './selectors'
+export * from './pagination'
+export * from './search'
+export * from './post'
